# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TEST",
        "title": "API",
        "contact": {},
        "version": "1.7"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/medals/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This endpoint adds a new medal.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Medal"
                ],
                "summary": "Add a medal",
                "parameters": [
                    {
                        "description": "Medal details to add",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.Medal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.Medal"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.Message"
                        }
                    }
                }
            }
        },
        "/medals/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This endpoint deletes a medal by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Medal"
                ],
                "summary": "Delete a medal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medal ID to delete",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.Message"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.Message"
                        }
                    }
                }
            }
        },
        "/medals/edit": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This endpoint edits an existing medal.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Medal"
                ],
                "summary": "Edit a medal",
                "parameters": [
                    {
                        "description": "Medal details to edit",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.Medal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.Medal"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.Message"
                        }
                    }
                }
            }
        },
        "/medals/get": {
            "get": {
                "description": "This endpoint retrieves a medal by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Medal"
                ],
                "summary": "Get a medal",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medal ID to retrieve",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.Medal"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.Message"
                        }
                    }
                }
            }
        },
        "/medals/getall": {
            "get": {
                "description": "This endpoint retrieves all medals with pagination and optional filters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Medal"
                ],
                "summary": "List medals",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Country ID",
                        "name": "country",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Event ID",
                        "name": "event_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Athlete ID",
                        "name": "athlete_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.ListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.Message"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.Message"
                        }
                    }
                }
            }
        },
        "/medals/ranking": {
            "get": {
                "description": "This endpoint retrieves the ranking of countries based on medals.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Medal"
                ],
                "summary": "Get medal rankings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.MedalRankingResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.Message"
                        }
                    }
                }
            }
        },
        "/stream/send": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This endpoint sends an event to the streaming service.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Live Streaming"
                ],
                "summary": "Send an event to the streaming service",
                "parameters": [
                    {
                        "description": "Event details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_stream_service.StreamEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_stream_service.StreamEventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_stream_service.StreamEventResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/olympy_api-gateway_genproto_stream_service.StreamEventResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "olympy_api-gateway_genproto_medal_service.CountryMedalCount": {
            "type": "object",
            "properties": {
                "bronze": {
                    "type": "integer"
                },
                "country_id": {
                    "type": "integer"
                },
                "country_name": {
                    "type": "string"
                },
                "gold": {
                    "type": "integer"
                },
                "ranking": {
                    "type": "integer"
                },
                "silver": {
                    "type": "integer"
                }
            }
        },
        "olympy_api-gateway_genproto_medal_service.ListResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "medals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.Medal"
                    }
                }
            }
        },
        "olympy_api-gateway_genproto_medal_service.Medal": {
            "type": "object",
            "properties": {
                "athlete_id": {
                    "type": "string"
                },
                "country_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "event_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "olympy_api-gateway_genproto_medal_service.MedalRankingResponse": {
            "type": "object",
            "properties": {
                "country_medal_counts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/olympy_api-gateway_genproto_medal_service.CountryMedalCount"
                    }
                }
            }
        },
        "olympy_api-gateway_genproto_medal_service.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "olympy_api-gateway_genproto_stream_service.StreamEventRequest": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "olympy_api-gateway_genproto_stream_service.StreamEventResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}